refactor: move material module inside shared module
